% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt_dials.R
\docType{data}
\name{dwt_coefs}
\alias{dwt_coefs}
\alias{values_dwt_coefs}
\title{DWT Coefficient Methods}
\format{
An object of class \code{character} of length 3.
}
\usage{
dwt_coefs(values = values_dwt_coefs)

values_dwt_coefs
}
\arguments{
\item{values}{A character string of possible values. See \code{values_prune_method}
in examples below.}
}
\description{
DWT Coefficient Methods
}
\details{
This parameter is used in \code{stepdwt::step_dwt}.
}
\examples{
values_dwt_coefs
dwt_coefs()
}
\keyword{datasets}
