% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt_dials.R
\docType{data}
\name{dwt_filter}
\alias{dwt_filter}
\alias{values_dwt_filter}
\title{DWT Filter Methods}
\format{
An object of class \code{character} of length 26.
}
\usage{
dwt_filter(values = values_dwt_filter)

values_dwt_filter
}
\arguments{
\item{values}{A character string of possible values. See \code{values_prune_method}
in examples below.}
}
\description{
DWT Filter Methods
}
\details{
This parameter is used in \code{stepdwt::step_dwt}.
}
\examples{
values_dwt_filter
dwt_filter()
}
\keyword{datasets}
